Moved module code into iotdevice, using embedded structs
