+++ common/tls.go
@@ -90,44 +90,42 @@ func RootCAs() *x509.CertPool {
 	return p
 }
 
+// TrustBundleResponse aids parsing the response from the edge.
 type TrustBundleResponse struct {
 	Certificate string `json:"certificate"`
 }
 
 // TrustBundle root CA certificates pool for connecting to EdgeHub Gateway.
-//
-// The TrustBundle is generated by the Edge Runtime at startup. The connection
-// to the edgeHub (proxy for IoT Hub) requires this cert. This CA cert can be
-// pulled from an API exposed on the edge runtime detailed in the below function.
-func TrustBundle(workloadUri, apiVersion string) *x509.CertPool {
+func TrustBundle(workloadURI string) (*x509.CertPool, error) {
 
 	// format uri string
-	uri := fmt.Sprintf("%strust-bundle?api-version=%s", workloadUri, apiVersion)
+	uri := fmt.Sprintf("%strust-bundle?api-version=2019-11-05", workloadURI)
 
 	// get http response and handle error
 	resp, err := http.Get(uri)
 	if err != nil {
-		panic("tls: unable to append certificates")
+		return nil, fmt.Errorf("tls: unable to append certificates: %s", err.Error())
 	}
 	defer resp.Body.Close()
 
 	// read response
 	body, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
-		panic("tls: unable to append certificates")
+		return nil, fmt.Errorf("tls: unable to append certificates: %s", err.Error())
 	}
 
 	// convert to struct
 	tbr := TrustBundleResponse{}
 	err = json.Unmarshal(body, &tbr)
 	if err != nil {
-		panic("tls: unable to append certificates")
+		return nil, fmt.Errorf("tls: unable to append certificates: %s", err.Error())
 	}
 
 	p := x509.NewCertPool()
-	if ok := p.AppendCertsFromPEM(caCerts); !ok {
-		panic("tls: unable to append certificates")
+	if ok := p.AppendCertsFromPEM([]byte(tbr.Certificate)); !ok {
+		err = fmt.Errorf("tls: unable to append certificates: count not append certs to pool")
 	}
-	return p
+
+	return p, err
 
 }
